refactor: fold the row-schema trait into RowsField and dedupe reserved-key plumbing
Simplify pass over the substrate:

- HandlesRowSchemas had one consumer and referenced RowsField back; its
  methods now live on RowsField, with a castRow() hook so subclasses
  decorate rows without repeating the guard-and-zip (TypedRowsField adds
  type there), and rulesForRows() so the type rules reuse the fetched rows.
- Rule::in() replaces the hand-concatenated in: rule and is built once per
  validate instead of per row; the type key is the TypedRowsField::TYPE
  constant, mirroring RowsField::ROW_ID.
- RowKeyInputs generalizes the hidden per-row inputs over the key, replacing
  the builder's two bespoke type-input sites and RowIdInputs.
- rowSchemaFor()/rowTemplatesOf() in row-templates.ts hold the typed-rows
  wire shape once for the builder and the prefill walker.
- The resolve/search transport no longer scrubs row ids: rowId is a real
  wire key the server validates, so resolver payloads match submits.
- registerRowCollectionType is dropped until a consumer exists, and the
  toolkit/API docblocks lose the marketing language.
