+++ resources/js/form/components/fields/repeater.tsx
@@ -4,7 +4,7 @@ import { FormFieldFrame } from "../base/field";
 import { useFormContext } from "../context";
 import { useDependentField } from "../use-dependent-field";
 import { ROW_ID_KEY } from "./repeater-rows";
-import { RowIdInputs } from "./row-id-inputs";
+import { RowKeyInputs } from "./row-key-inputs";
 import { RowItem } from "./row-item";
 import { columnsFromSchema, TableRows } from "./table-rows";
 import { useFlipReorder } from "./use-flip-reorder";
@@ -61,7 +61,7 @@ export const RepeaterComponent: RendererComponent<"field.repeater"> = ({ node })
       required={required}
     >
       <div className="flex flex-col gap-3">
-        <RowIdInputs path={path} rows={rows} />
+        <RowKeyInputs path={path} rows={rows} rowKey={ROW_ID_KEY} />
         {isTable ? (
           <TableRows
             base={path}
